# apps/docs/public/contracts/openapi-web.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "deviceName": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Devicename"
          },
          "password": {
            "format": "password",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "username": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "SessionUserResponse": {
        "additionalProperties": false,
        "properties": {
          "authorizedAccountIds": {
            "items": {
              "type": "string"
            },
            "title": "Authorizedaccountids",
            "type": "array"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "displayName",
          "permissions",
          "authorizedAccountIds"
        ],
        "title": "SessionUserResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebSessionGrantResponse": {
        "additionalProperties": false,
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "title": "Accesstokenexpiresat",
            "type": "string"
          },
          "deviceSessionId": {
            "title": "Devicesessionid",
            "type": "string"
          },
          "tokenType": {
            "default": "Bearer",
            "title": "Tokentype",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/SessionUserResponse"
          }
        },
        "required": [
          "accessToken",
          "accessTokenExpiresAt",
          "deviceSessionId",
          "user"
        ],
        "title": "WebSessionGrantResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "QuantX Web 会话与服务状态接口。Web 业务能力使用同源 /graphql；development 会话端点仅在开发环境可用。",
    "title": "QuantX Web API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/web/session": {
      "delete": {
        "description": "Idempotently revoke the current browser session and clear its cookie.",
        "operationId": "delete_web_session_auth_web_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Delete Web Session",
        "tags": [
          "authentication"
        ],
        "x-quantx-stability": "web-internal"
      },
      "post": {
        "description": "Create a browser session without exposing its refresh token to JavaScript.",
        "operationId": "create_web_session_auth_web_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebSessionGrantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Web Session",
        "tags": [
          "authentication"
        ],
        "x-quantx-stability": "web-internal"
      }
    },
    "/auth/web/session/development": {
      "post": {
        "description": "Create a passwordless browser session for a development-only DB user.",
        "operationId": "create_development_web_session_auth_web_session_development_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebSessionGrantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Development auto-login is disabled"
          },
          "503": {
            "description": "Development database user is not configured"
          }
        },
        "summary": "Create Development Web Session",
        "tags": [
          "authentication"
        ],
        "x-quantx-stability": "development-only"
      }
    },
    "/auth/web/session/refresh": {
      "post": {
        "description": "Rotate the HttpOnly refresh cookie and return a short-lived access token.",
        "operationId": "refresh_web_session_auth_web_session_refresh_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebSessionGrantResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Web session is missing or expired"
          }
        },
        "summary": "Refresh Web Session",
        "tags": [
          "authentication"
        ],
        "x-quantx-stability": "web-internal"
      }
    },
    "/health": {
      "get": {
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health Check",
        "x-quantx-stability": "web-internal"
      }
    },
    "/health/components": {
      "get": {
        "operationId": "health_components_health_components_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health Components",
        "x-quantx-stability": "web-internal"
      }
    },
    "/health/live": {
      "get": {
        "operationId": "health_live_health_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health Live",
        "x-quantx-stability": "web-internal"
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "health_ready_health_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health Ready",
        "x-quantx-stability": "web-internal"
      }
    }
  }
}